saving correct images, correct dtype for hough visualisation

+++ src/gui_pipeline.py
@@ -412,7 +412,7 @@ class Hough(Tab):
         self._draw_lines(tgt)
         self._draw_points(tgt)
         self._draw_triangles(tgt)
-        return tgt
+        return tgt.astype(np.uint8)
 
     def _draw_result(self):
         tgt = self._mod_hough.arr / 5
